+++ formatter.go
@@ -25,9 +25,9 @@ type TextFormatter struct {
 	names  []string
 }
 
-// NewTextFormatter creates a textFormatter which uses LogRecord attributes to
-// contribute logging string, e.g. %(message)s or %(levelno)d. See LogRecord for
-// more details.
+// NewTextFormatter creates a textFormatter which uses LogRecord macros and
+// format string to contribute logging string, e.g. %(message)s or %(levelno)d.
+// See LogRecord for more details.
 func NewTextFormatter(s string) TextFormatter {
 	var tf = TextFormatter{}
 	var i, n = 0, len(s)
@@ -85,8 +85,7 @@ func (tf TextFormatter) Format(record LogRecord) (string, error) {
 	return fmt.Sprintf(tf.fmtstr, attrs...), nil
 }
 
-// JSONFormatter allows logging message to be parsed as json format. It uses the
-// same macros as other Formatter.
+// JSONFormatter allows logging message to be parsed as json format.
 type JSONFormatter struct {
 	fields []field
 }
@@ -96,27 +95,88 @@ func NewJSONFormatter() *JSONFormatter {
 	return &JSONFormatter{fields: make([]field, 0, 5)}
 }
 
-// AddField adds the macro to logging message under key name. It returns itself.
+// AddField adds the macro value to the logging message under a name. It returns
+// itself. For {message} macro and JSON event logger, you could leave the name
+// as empty if you want to add all fields of message into the outer object
+// directly.
 func (js *JSONFormatter) AddField(name, macro string) *JSONFormatter {
 	js.fields = append(js.fields, field{key: name, value: macro})
 	return js
 }
 
-// Format creates the logging message as the json format.
+// Format creates the logging message of JSON format.
 func (js JSONFormatter) Format(record LogRecord) (string, error) {
 	var err error
-	var attr = make(map[string]any)
+	var data = make(map[string]any)
 	for _, f := range js.fields {
-		attr[f.key], err = record.getAttributeByName(f.value.(string))
+		var attr, err = record.getAttributeByName(f.value.(string))
 		if err != nil {
 			return "", err
+		}
+
+		if mattr, ok := attr.(map[string]any); ok && f.key == "" {
+			for k, v := range mattr {
+				data[k] = v
+			}
+		} else {
+			data[f.key] = attr
 		}
 	}
 
 	var s []byte
-	s, err = json.Marshal(attr)
+	s, err = json.Marshal(data)
 	if err != nil {
 		return "", xyerror.ValueError.New(err)
 	}
 	return string(s), nil
+}
+
+// StructureFormatter formats the logging message with the form of key=value.
+type StructureFormatter struct {
+	fields []field
+}
+
+// NewStructureFormatter creates an empty StructureFormatter.
+func NewStructureFormatter() *StructureFormatter {
+	return &StructureFormatter{}
+}
+
+// AddField adds the macro to logging message under a name. It returns itself.
+// If you leave the name as empty, it will adds the macro value without the name
+// and equal character.
+func (sf *StructureFormatter) AddField(name, macro string) *StructureFormatter {
+	sf.fields = append(sf.fields, field{key: name, value: macro})
+	return sf
+}
+
+// Format creates the logging message with the form of key=value.
+func (sf StructureFormatter) Format(record LogRecord) (string, error) {
+	var msg string
+	for _, f := range sf.fields {
+		var attr, err = record.getAttributeByName(f.value.(string))
+		if err != nil {
+			return "", err
+		}
+
+		var value string
+		switch attr.(type) {
+		case map[string]any:
+			var s []byte
+			s, err = json.Marshal(attr)
+			if err != nil {
+				return "", err
+			}
+			value = string(s)
+		default:
+			value = fmt.Sprint(attr)
+		}
+
+		if f.key == "" {
+			msg = prefixMessage(msg, value)
+		} else {
+			msg = prefixMessage(msg, fmt.Sprintf("%s=%s", f.key, value))
+		}
+	}
+
+	return msg, nil
 }